fix(run): avoid no-op false positives for reasoning output
Treat high reported output tokens as meaningful reasoning work for single-turn zero-tool sessions. Carry transport output token metadata into post-exec and add a regression for security-audit style reasoning-only output.

Split post-exec/session-exec helpers and no-op gate tests while touching these paths so monolith guards keep passing; bump workspace version to 0.1.799.
