+++ crates/cli-sub-agent/src/pipeline_tests_no_op_gate.rs
@@ -36,6 +36,7 @@ fn build_test_ctx<'a>(
         pre_exec_snapshot: None,
         has_tool_calls,
         turn_count: 0,
+        output_tokens: None,
         sa_mode,
     }
 }
@@ -66,6 +67,9 @@ fn write_result_sidecar(session_dir: &std::path::Path, contents: &str) {
         .expect("create output dir");
     std::fs::write(path, contents).expect("write output/result.toml");
 }
+
+#[path = "pipeline_tests_no_op_gate_task_type.rs"]
+mod task_type_tests;
 
 #[tokio::test]
 async fn permanent_tool_quota_sets_tool_exhausted_phase() {
@@ -389,6 +393,52 @@ async fn no_op_gate_does_not_trigger_when_elapsed_exceeds_threshold() {
 }
 
 #[tokio::test]
+async fn no_op_gate_does_not_trigger_with_high_output_tokens() {
+    let tmp = tempfile::tempdir().expect("tempdir");
+    let _sandbox = ScopedSessionSandbox::new(&tmp).await;
+    let project_root = tmp.path();
+    let mut session =
+        create_session(project_root, Some("test"), None, Some("claude-code")).expect("create");
+    let session_dir =
+        csa_session::get_session_dir(project_root, &session.meta_session_id).expect("dir");
+
+    let executor = Executor::ClaudeCode {
+        model_override: None,
+        thinking_budget: None,
+        runtime_metadata: ClaudeCodeRuntimeMetadata::current(),
+    };
+    let hooks_config = csa_hooks::HooksConfig::default();
+    let start = chrono::Utc::now() - chrono::Duration::seconds(15);
+    let mut ctx = build_test_ctx(
+        &executor,
+        session_dir,
+        project_root,
+        start,
+        &hooks_config,
+        false,
+        true,
+    );
+    ctx.output_tokens = Some(14_093);
+    let mut result = build_test_result("Security audit completed with a PASS verdict.");
+
+    process_execution_result(ctx, &mut session, &mut result)
+        .await
+        .expect("process_execution_result");
+
+    let persisted = load_result(project_root, &session.meta_session_id)
+        .expect("load")
+        .expect("result exists");
+    assert_eq!(persisted.exit_code, 0);
+    assert_eq!(persisted.status, SessionResult::status_from_exit_code(0));
+    assert!(
+        !persisted.summary.starts_with("no-op exit detected"),
+        "high output_tokens must suppress no-op rewrite, got: {}",
+        persisted.summary
+    );
+    assert_eq!(result.exit_code, 0);
+}
+
+#[tokio::test]
 async fn no_op_gate_preserves_original_summary_as_suffix() {
     let tmp = tempfile::tempdir().expect("tempdir");
     let _sandbox = ScopedSessionSandbox::new(&tmp).await;
@@ -585,160 +635,5 @@ async fn no_op_gate_does_not_trigger_when_turn_count_exceeds_one() {
     assert!(
         !persisted.summary.starts_with("no-op exit detected"),
         "gate must not fire for multi-turn sessions"
-    );
-}
-
-#[tokio::test]
-async fn no_op_gate_does_not_trigger_for_non_run_task_type() {
-    for task_type in &["review", "debate", "plan"] {
-        let tmp = tempfile::tempdir().expect("tempdir");
-        let _sandbox = ScopedSessionSandbox::new(&tmp).await;
-        let project_root = tmp.path();
-        let mut session =
-            create_session(project_root, Some("test"), None, Some("claude-code")).expect("create");
-        let session_dir =
-            csa_session::get_session_dir(project_root, &session.meta_session_id).expect("dir");
-
-        let executor = Executor::ClaudeCode {
-            model_override: None,
-            thinking_budget: None,
-            runtime_metadata: ClaudeCodeRuntimeMetadata::current(),
-        };
-        let hooks_config = csa_hooks::HooksConfig::default();
-        let start = chrono::Utc::now() - chrono::Duration::seconds(15);
-        let mut ctx = build_test_ctx(
-            &executor,
-            session_dir,
-            project_root,
-            start,
-            &hooks_config,
-            false,
-            true,
-        );
-        ctx.task_type = Some(task_type);
-        let mut result = build_test_result("Review completed successfully.");
-
-        process_execution_result(ctx, &mut session, &mut result)
-            .await
-            .expect("process_execution_result");
-
-        let persisted = load_result(project_root, &session.meta_session_id)
-            .expect("load")
-            .expect("result exists");
-        assert_eq!(
-            persisted.exit_code, 0,
-            "gate must not fire for task_type={task_type}"
-        );
-        assert_eq!(
-            persisted.status,
-            SessionResult::status_from_exit_code(0),
-            "status must remain success for task_type={task_type}"
-        );
-        assert!(
-            !persisted.summary.starts_with("no-op exit detected"),
-            "summary must NOT be prefixed for task_type={task_type}, got: {}",
-            persisted.summary
-        );
-    }
-}
-
-#[tokio::test]
-async fn no_op_gate_still_triggers_for_run_task_type() {
-    let tmp = tempfile::tempdir().expect("tempdir");
-    let _sandbox = ScopedSessionSandbox::new(&tmp).await;
-    let project_root = tmp.path();
-    let mut session =
-        create_session(project_root, Some("test"), None, Some("claude-code")).expect("create");
-    let session_dir =
-        csa_session::get_session_dir(project_root, &session.meta_session_id).expect("dir");
-
-    let executor = Executor::ClaudeCode {
-        model_override: None,
-        thinking_budget: None,
-        runtime_metadata: ClaudeCodeRuntimeMetadata::current(),
-    };
-    let hooks_config = csa_hooks::HooksConfig::default();
-    let start = chrono::Utc::now() - chrono::Duration::seconds(10);
-    let mut ctx = build_test_ctx(
-        &executor,
-        session_dir,
-        project_root,
-        start,
-        &hooks_config,
-        false,
-        true,
-    );
-    ctx.task_type = Some("run");
-    let mut result = build_test_result("I'll start by exploring.");
-
-    process_execution_result(ctx, &mut session, &mut result)
-        .await
-        .expect("process_execution_result");
-
-    let persisted = load_result(project_root, &session.meta_session_id)
-        .expect("load")
-        .expect("result exists");
-    assert_eq!(persisted.exit_code, 1, "gate must fire for task_type=run");
-    assert_eq!(
-        persisted.status,
-        SessionResult::status_from_exit_code(1),
-        "status must be failure for task_type=run"
-    );
-    assert!(
-        persisted.summary.starts_with("no-op exit detected"),
-        "summary must be prefixed for task_type=run, got: {}",
-        persisted.summary
-    );
-}
-
-#[tokio::test]
-async fn no_op_gate_triggers_when_task_type_is_none() {
-    let tmp = tempfile::tempdir().expect("tempdir");
-    let _sandbox = ScopedSessionSandbox::new(&tmp).await;
-    let project_root = tmp.path();
-    let mut session =
-        create_session(project_root, Some("test"), None, Some("claude-code")).expect("create");
-    let session_dir =
-        csa_session::get_session_dir(project_root, &session.meta_session_id).expect("dir");
-
-    let executor = Executor::ClaudeCode {
-        model_override: None,
-        thinking_budget: None,
-        runtime_metadata: ClaudeCodeRuntimeMetadata::current(),
-    };
-    let hooks_config = csa_hooks::HooksConfig::default();
-    let start = chrono::Utc::now() - chrono::Duration::seconds(10);
-    let mut ctx = build_test_ctx(
-        &executor,
-        session_dir,
-        project_root,
-        start,
-        &hooks_config,
-        false,
-        true,
-    );
-    ctx.task_type = None;
-    let mut result = build_test_result("I'll start by exploring.");
-
-    process_execution_result(ctx, &mut session, &mut result)
-        .await
-        .expect("process_execution_result");
-
-    let persisted = load_result(project_root, &session.meta_session_id)
-        .expect("load")
-        .expect("result exists");
-    assert_eq!(
-        persisted.exit_code, 1,
-        "gate must fire when task_type is None"
-    );
-    assert_eq!(
-        persisted.status,
-        SessionResult::status_from_exit_code(1),
-        "status must be failure when task_type is None"
-    );
-    assert!(
-        persisted.summary.starts_with("no-op exit detected"),
-        "summary must be prefixed when task_type is None, got: {}",
-        persisted.summary
     );
 }
